Add OBOE audio architecture. Set version to 2.21.4.

+++ architecture/faust/audio/android-dsp.h
@@ -39,7 +39,7 @@
 #define NUM_OUTPUTS 2
 #define CPU_TABLE_SIZE 16
 
-static const char *res_str(SLresult result)
+static const char* res_str(SLresult result)
 {
     switch(result)
     {
@@ -127,7 +127,7 @@ class androidaudio : public audio {
         unsigned int fBufferSize;
     
         int64_t fCPUTable[CPU_TABLE_SIZE];
-        int fCPUTableIndex;
+        int64_t fCPUTableIndex;
     
         float** fInputs;
         float** fOutputs;
